refactor : facke 및 빌더 분리
